Move file metadata interactor functions to js modules
